// src/app/shared/model/cat.ts
export interface Cat {
    id: number;
    name: string;
    sex: string;
    age: number;
    breed: string;
    character: string;
    description: string;
    urgent: boolean;
    user_id: number;
    localisation_id: number;
    localisation: string;
    fav : boolean;
    photo: string;
    count: number;
    statut: number;
}